+++ src/main/java/io/github/astrapi69/entity/activatable/ActivatableEntity.java
@@ -31,9 +31,11 @@ import javax.persistence.AccessType;
 import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
 
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 import lombok.experimental.SuperBuilder;
 import io.github.astrapi69.data.activatable.IdentifiableActivatable;
 import io.github.astrapi69.entity.base.SequenceBaseEntity;
@@ -46,6 +48,8 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
  *            the generic type of the id
  */
 @MappedSuperclass
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @Access(AccessType.FIELD)
 @Getter
 @Setter