+++ src/main/java/io/github/astrapi69/entity/modifiable/ByLastModification.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -50,10 +51,11 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
  *            the generic type of the user or account
  */
 @MappedSuperclass
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @Entity
 @Getter
 @Setter
-@ToString
 @NoArgsConstructor
 @SuperBuilder
 public abstract class ByLastModification<PK extends Serializable, T, U>