+++ src/main/java/io/github/astrapi69/entity/modifiable/LastModification.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -48,10 +49,11 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
  *            the generic type of time measurement
  */
 @MappedSuperclass
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @Entity
 @Getter
 @Setter
-@ToString
 @NoArgsConstructor
 @SuperBuilder
 public abstract class LastModification<PK extends Serializable, T> extends SequenceBaseEntity<PK>