+++ src/main/java/io/github/astrapi69/entity/treeable/TreeUUIDEntity.java
@@ -32,9 +32,11 @@ import javax.persistence.ManyToOne;
 import javax.persistence.MappedSuperclass;
 
 import lombok.AccessLevel;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 import io.github.astrapi69.data.treeable.Treeable;
@@ -45,6 +47,8 @@ import io.github.astrapi69.entity.uniqueable.UUIDEntity;
  * {@link TreeUUIDEntity} has no parent, all other {@link TreeUUIDEntity} objects have a parent.
  */
 @MappedSuperclass
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @Getter
 @Setter
 @NoArgsConstructor