+++ src/main/java/io/github/astrapi69/entity/validatable/ValidatableEntity.java
@@ -29,6 +29,7 @@ import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
 
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -45,10 +46,11 @@ import io.github.astrapi69.entity.base.SequenceBaseEntity;
  *            the generic type of the id
  */
 @MappedSuperclass
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @Entity
 @Getter
 @Setter
-@ToString
 @NoArgsConstructor
 @SuperBuilder
 public abstract class ValidatableEntity<PK extends Serializable, T> extends SequenceBaseEntity<PK>